requirements.txt
flask==2.2.3
python-binance==1.0.16
python-dotenv==1.0.0
flask-wtf==1.1.1
requests==2.28.2